requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic
pydantic-settings
websockets
redis
httpx
pillow
email-validator
pytest
pytest-asyncio
httpx
faker